Initial.
This is the initial structure for replicating the domain-driven
design book example using the Ruby on Rails framework along with a
hexagonal architecture.

Functionality has been partitioned into gems in order to enforce
separation of concerns.  An alternative approach would be to just
be disciplined with the Rails app itself.

Were we to productionize this project, the gems would be separated
out.